+++ CHANGES.rst
@@ -4,7 +4,9 @@ Changelog
 0.2.4
 ~~~~~
 
-* using mixin approach for RoleFilterModelViewSet: code moved to RoleFilterMixin
+* Using mixin approach for RoleFilterModelViewSet: code moved to RoleFilterMixin.
+* get_allowed_actions now also accepts :code:`obj` argument that is used to verify actions
+  allowed during :code:`check_object_permissions`.
 
 0.2.3
 ~~~~~